split wfstable streams to easier testing
