+++ web/client/components/widgets/enhancers/wfsTable/triggerFetch.js
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2018, GeoSolutions Sas.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+require('rxjs');
+const {getLayerUrl} = require('./common');
+const sameFilter = (f1, f2) => f1 === f2;
+const sameOptions = (o1 = {}, o2 = {}) =>
+    o1.propertyName === o2.propertyName;
+const sameSortOptions = (o1 = {}, o2 = {}) =>
+    o1.sortBy === o2.sortBy
+    && o1.sortOrder === o2.sortOrder;
+
+/**
+ * Function that converts stream of a WFSTable props to trigger data fetch events
+ * @param {Observable} Stream of props.
+ * @return {Observable} Stream of props to trigger the data fetch
+ */
+module.exports = ($props) =>
+    $props.filter(({ layer = {}, options }) => layer.name && options)
+        .distinctUntilChanged(
+        ({ layer = {}, options = {}, filter, sortOptions }, newProps) =>
+            getLayerUrl(layer) === getLayerUrl(layer)
+            && (newProps.layer && layer.name === newProps.layer.name)
+            && sameOptions(options, newProps.options)
+            && sameFilter(filter, newProps.filter)
+            && sameSortOptions(sortOptions, newProps.sortOptions))
+        // when one of the items above changed invalidates cache for before the next request
+        .map((props) => ({
+            ...props,
+            features: [],
+            pages: [],
+            pagination: {}
+        }));
